Select mp4/mp3 added; Icon added; "imagens" folder added

+++ download.py
@@ -3,7 +3,8 @@ from pytube import YouTube as Yt
 from moviepy.editor import *
 import os
 
-def baixar(id, title, labelmessage):
+def baixar(id, title, labelmessage, cb):
+            cbtt = cb.get()
             code = id.get()
             filetitle = str(title.get())
             dto = askdirectory()
@@ -18,13 +19,15 @@ def baixar(id, title, labelmessage):
             audio_video = video_mp4.streams.get_audio_only()
             
             try:
-
                 midea_dir = dto +  f"/{filetitle}"
-                audio_video.download(output_path=f'{dto}', filename=f'{filetitle}.mp4')
-                video = AudioFileClip(midea_dir + '.mp4')
-                video.write_audiofile(midea_dir +'.mp3')
-                os.remove(midea_dir + '.mp4')
-                os.system("cls")
+                if cbtt:
+                    baixar = audio_video.download(output_path=f'{dto}', filename=f'{filetitle}.mp4')
+                else:
+                    baixar
+                    video = AudioFileClip(midea_dir + '.mp4')
+                    video.write_audiofile(midea_dir +'.mp3')
+                    os.remove(midea_dir + '.mp4')
+                    os.system("cls")
 
             except Exception as e:
                 labelmessage['text'] = "Algo está errado, cheque o ID/link do vídeo"
@@ -33,4 +36,4 @@ def baixar(id, title, labelmessage):
             else: 
                 labelmessage["text"] = "Verifique o arquivo no diretório combinado"
                 id.delete(0, "end")
-
+                title.delete(0, "end")
